requirements.txt
ccxt
pandas
pyyaml
python-dotenv
tenacity
streamlit
plotly